+++ source/extensions/isaacsim.robot.schema/robot_schema/__init__.py
@@ -37,7 +37,7 @@ class Attributes(Enum):
     TR_X_OFFSET = (f"{_attr_prefix}:physics:Tr_X:DofOffset", "Tr X Offset", pxr.Sdf.ValueTypeNames.Int)
     TR_Y_OFFSET = (f"{_attr_prefix}:physics:Tr_Y:DofOffset", "Tr Y Offset", pxr.Sdf.ValueTypeNames.Int)
     TR_Z_OFFSET = (f"{_attr_prefix}:physics:Tr_Z:DofOffset", "Tr Z Offset", pxr.Sdf.ValueTypeNames.Int)
-    RETRY_INTERVAL = (f"{_attr_prefix}:retryInterval", "Retry Interval", pxr.Sdf.ValueTypeNames.Double)
+    RETRY_INTERVAL = (f"{_attr_prefix}:retryInterval", "Retry Interval", pxr.Sdf.ValueTypeNames.Float)
     ACCELERATION_LIMIT = (
         f"{_attr_prefix}:physics:AccelerationLimit",
         "Acceleration Limit",
@@ -46,11 +46,10 @@ class Attributes(Enum):
     JERK_LIMIT = (f"{_attr_prefix}:physics:JerkLimit", "Jerk Limit", pxr.Sdf.ValueTypeNames.FloatArray)
     ACTUATOR = (f"{_attr_prefix}:physics:Actuator", "Actuator", pxr.Sdf.ValueTypeNames.BoolArray)
     STATUS = (f"{_attr_prefix}:status", "Status", pxr.Sdf.ValueTypeNames.Token)
-    SHEAR_FORCE_LIMIT = (f"{_attr_prefix}:shearForceLimit", "Shear Force Limit", pxr.Sdf.ValueTypeNames.Double)
-    COAXIAL_FORCE_LIMIT = (f"{_attr_prefix}:coaxialForceLimit", "Coaxial Force Limit", pxr.Sdf.ValueTypeNames.Double)
-    MAX_GRIP_DISTANCE = (f"{_attr_prefix}:maxGripDistance", "Max Grip Distance", pxr.Sdf.ValueTypeNames.Double)
-    GRIP_DISTANCE = (f"{_attr_prefix}:gripDistance", "Grip Distance", pxr.Sdf.ValueTypeNames.Double)
-    CLEARANCE_OFFSET = (f"{_attr_prefix}:clearanceOffset", "Clearance Offset", pxr.Sdf.ValueTypeNames.Double)
+    SHEAR_FORCE_LIMIT = (f"{_attr_prefix}:shearForceLimit", "Shear Force Limit", pxr.Sdf.ValueTypeNames.Float)
+    COAXIAL_FORCE_LIMIT = (f"{_attr_prefix}:coaxialForceLimit", "Coaxial Force Limit", pxr.Sdf.ValueTypeNames.Float)
+    MAX_GRIP_DISTANCE = (f"{_attr_prefix}:maxGripDistance", "Max Grip Distance", pxr.Sdf.ValueTypeNames.Float)
+    CLEARANCE_OFFSET = (f"{_attr_prefix}:clearanceOffset", "Clearance Offset", pxr.Sdf.ValueTypeNames.Float)
 
     # Custom properties for name and type
     @property
@@ -68,9 +67,9 @@ class Attributes(Enum):
 
 class Relations(Enum):
     ROBOT_LINKS = (f"{_attr_prefix}:physics:robotLinks", "Robot Links")
-    ROBOT_JOINTS = (f"{_attr_prefix}:physics:robotjoints", "Robot Joints")
-    ATTACHMENT_POINTS = ("attachmentPoints", "Attachment Points")
-    GRIPPED_OBJECTS = ("grippedObjects", "Gripped Objects")
+    ROBOT_JOINTS = (f"{_attr_prefix}:physics:robotJoints", "Robot Joints")
+    ATTACHMENT_POINTS = (f"{_attr_prefix}:attachmentPoints", "Attachment Points")
+    GRIPPED_OBJECTS = (f"{_attr_prefix}:grippedObjects", "Gripped Objects")
 
     @property
     def name(self):
@@ -87,7 +86,7 @@ def ApplyRobotAPI(prim: pxr.Usd.Prim):
         prim.CreateAttribute(attr.name, attr.type, True)
 
     for rel in [Relations.ROBOT_LINKS, Relations.ROBOT_JOINTS]:
-        prim.CreateRelationship(rel.value, custom=True)
+        prim.CreateRelationship(rel.name, custom=True)
 
 
 def ApplyLinkAPI(prim: pxr.Usd.Prim):
@@ -137,13 +136,13 @@ def CreateSurfaceGripper(stage: pxr.Usd.Stage, prim_path: str) -> pxr.Usd.Prim:
         Attributes.SHEAR_FORCE_LIMIT,
         Attributes.COAXIAL_FORCE_LIMIT,
         Attributes.MAX_GRIP_DISTANCE,
-        Attributes.GRIP_DISTANCE,
+        Attributes.RETRY_INTERVAL,
     ]:
-        prim.CreateAttribute(attr.name, attr.type, True)
+        prim.CreateAttribute(attr.name, attr.type, False)
 
     # Create relationships
-    prim.CreateRelationship(Relations.ATTACHMENT_POINTS.name, custom=True)
-    prim.CreateRelationship(Relations.GRIPPED_OBJECTS.name, custom=True)
+    prim.CreateRelationship(Relations.ATTACHMENT_POINTS.name, False)
+    prim.CreateRelationship(Relations.GRIPPED_OBJECTS.name, False)
 
     return prim
 
@@ -151,4 +150,4 @@ def CreateSurfaceGripper(stage: pxr.Usd.Stage, prim_path: str) -> pxr.Usd.Prim:
 def ApplyAttachmentPointAPI(prim: pxr.Usd.Prim):
     prim.AddAppliedSchema(Classes.ATTACHMENT_POINT_API.value)
     for attr in [Attributes.FORWARD_AXIS, Attributes.CLEARANCE_OFFSET]:
-        prim.CreateAttribute(attr.name, attr.type, True)
+        prim.CreateAttribute(attr.name, attr.type, False)