+++ source/extensions/isaacsim.robot.surface_gripper/python/__init__.py
@@ -9,5 +9,4 @@
 # its affiliates is strictly prohibited.
 
 # there is no extension, so we just import bindings
-from . import _surface_gripper
 from .impl import *